fix(e2e): stabilize homepage network filter Appium smoke on iOS
Use iOS wallet readiness in waitForAppReady, harden missing-element assertions,
and align the homepage filter spec with reliable wallet-home waits so login and
token visibility checks pass consistently on Appium.
